Adding more new stuff and updating others

+++ backend/src/main/java/com/moviedates/backend/model/Vote.java
@@ -22,7 +22,7 @@ public class Vote {
     @JoinColumn(name = "session_id")
     private Session session;
 
-    private Integer movieId; // The ID from TMDB API
+    private long movieId; // The ID from TMDB API
 
     private boolean accepted; // true = Right Swipe (Accept), false = Left Swipe (Reject)
